@@ -45,7 +45,7 @@ class UserAuthenticator extends AbstractLoginFormAuthenticator
          return new RedirectResponse($this->urlGenerator->generate('task_list'));
     }
 
-    protected function getLoginUrl(Request $request): string
+    public function getLoginUrl(Request $request): string
     {
         return $this->urlGenerator->generate(self::LOGIN_ROUTE);
     }
