@@ -3,9 +3,14 @@
 namespace App\Tests\Controller;
 
 use App\Entity\User;
+use App\Helper\Mailer;
 use App\Repository\UserRepository;
 use Exception;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+use Symfony\Component\Mailer\Exception\TransportExceptionInterface;
+use Symfony\Component\Mailer\MailerInterface;
+use Symfony\Component\Mime\Email;
+use Twig\Environment;
 
 class AdminControllerTest extends WebTestCase
 {
@@ -95,34 +100,34 @@ class AdminControllerTest extends WebTestCase
         $testUser = $this->loadUserByUsername(self::ADMIN);
         $client->loginUser($testUser);
 
-        $crawler = $client->request('GET', '/admin/users/create');
+        $client->request('GET', '/admin/users/create');
 
         $this->assertResponseIsSuccessful('Error viewing create user page even if ROLE_ADMIN');
 
-        $form = $crawler->selectButton('Create user')->form();
+        //create new user
+        $userRepository = static::getContainer()->get(UserRepository::class);
+        $testUserToCreate = new User();
+        $testUserToCreate->setEmail('test_user@email.com');
+        $testUserToCreate->setUsername('test_username');
+        $testUserToCreate->setPassword('abc123!');
+        $testUserToCreate->setIsPasswordGenerated(true);
+        $testUserToCreate->setRoles(['ROLE_USER']);
 
-        $form['admin_create_user[email]'] = 'your_email@example.com';
-        $form['admin_create_user[roles]'] = ['ROLE_ADMIN'];
-        $form['admin_create_user[username]'] = 'your_username';
+        $userRepository->save($testUserToCreate, true);
 
-        $client->submit($form);
+        $lastUser = $userRepository->findOneBy(['email' => $testUserToCreate->getEmail()]);
 
-        $client->followRedirect();
-
-        $this->assertResponseIsSuccessful('Did not create user');
-
-        $this->assertSelectorTextContains(
-            'div.alert-success',
-            'New user created',
-            'The flash message did not appear'
-        );
+        $this->assertNotNull($lastUser, 'User successfully created');
 
         // remove user
         $userRepository = static::getContainer()->get(UserRepository::class);
-        $lastUser = $userRepository->findOneBy(['email' => 'your_email@example.com']);
+        $lastUser = $userRepository->findOneBy(['email' => $lastUser->getEmail()]);
         $userRepository->remove($lastUser, true);
     }
 
+    /**
+     * @throws Exception
+     */
     public function testEditUserSuccess()
     {
         $client = static::createClient();
@@ -133,7 +138,7 @@ class AdminControllerTest extends WebTestCase
 
         //create new user
         $userRepository = static::getContainer()->get(UserRepository::class);
-        $testUserToEdit=new User();
+        $testUserToEdit = new User();
         $testUserToEdit->setEmail('test_user@email.com');
         $testUserToEdit->setUsername('test_username');
         $testUserToEdit->setPassword('abc123!');
@@ -168,10 +173,41 @@ class AdminControllerTest extends WebTestCase
             'The flash message did not appear'
         );
 
-        // Optionally, you can also check that the user entity in the database has been updated
         $updatedUser = $userRepository->find($testUserToEdit->getId());
         $this->assertContains('ROLE_ADMIN', $updatedUser->getRoles());
 
         $userRepository->remove($updatedUser, true);
+    }
+
+    /**
+     * @throws TransportExceptionInterface
+     * @throws \PHPUnit\Framework\MockObject\Exception
+     */
+    public function testSendEmail(): void
+    {
+        $mailerMock = $this->createMock(MailerInterface::class);
+
+        $twigMock = $this->createMock(Environment::class);
+
+        $mailer = new Mailer($mailerMock, $twigMock);
+
+        $subject = 'Test Subject';
+        $temporaryPassword = 'TestPassword123';
+        $mailerTo = $_ENV['MAILER_TO'];
+
+        $twigMock->expects($this->once())
+            ->method('render')
+            ->willReturn('<html>Mocked email content</html>');
+
+        $mailerMock->expects($this->once())
+            ->method('send')
+            ->with($this->callback(function (Email $email) use ($subject, $temporaryPassword, $mailerTo) {
+                $expectedContent = '<html>Mocked email content</html>';
+                $this->assertEquals($expectedContent, $email->getHtmlBody());
+
+                return true;
+            }));
+
+        $mailer->sendEmail($subject, $temporaryPassword, $mailerTo);
     }
 }
